# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/in-memory": {
            "get": {
                "description": "list in memory records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inMemory"
                ],
                "summary": "list in memory records",
                "parameters": [
                    {
                        "type": "string",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "create in memory record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inMemory"
                ],
                "summary": "create in memory record",
                "parameters": [
                    {
                        "description": "in memory info",
                        "name": "tasks",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.InMemoryPostRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "delete in memory records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inMemory"
                ],
                "summary": "delete in memory records",
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.InMemoryPostRequest": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}